Add SELFMemory management module
* This new module creates classes for real, int32, and int64
multi-dimensional arrays that also have c_ptr counterparts for device
storage. This cuts down on the number of lines of code for higher level
modules that have classes that require both host and device storage.
Changes to incorporate this new memory management module have been
carried through to the NodalSEMData.F90 routines. We now need to work
on the SEMMesh and MappedNodalSEMData