Add additionaloutput call
This call is useful for appending additional attributes in child models
(like the Euler2D model). The Euler2D model has `primitive` and
`diagnostics` attributes that contain useful data for visualization.
This new function allows developers to add additional data (on top of
the solution attribute) to the hdf5 output files.